Number Formatter

Formats a numeric vector in a way that its largest absolute value determines the number of digits after the decimal separator. This function is helpful in perfectly aligning numbers on plots. Does not use scientific formatting.

format_max(x, digits = 4L, ...)

Arguments

x

A numeric vector to be formatted.

digits

Number of significant digits of the largest absolute value.

...

Further arguments passed to format(), e.g., big.mark = "'".

Value

A character vector of formatted numbers.

Examples

x <- c(100, 1, 0.1)
format_max(x)
#> [1] "100.0" "1.0"   "0.1"  

y <- c(100, 1.01)
format_max(y)
#> [1] "100" "1"  
format_max(y, digits = 5)
#> [1] "100.00" "1.01"